requirements.txt
flask
flask-cors
tensorflow==2.15.0
pillow
numpy
gunicorn
requests
